drf-next-project/django-drf/pyproject.toml
[tool.poetry]
name = "drf-nextjs"
version = "0.1.0"
description = ""
authors = ["rkhaq <rqk.email2@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2"
django-mptt = "^0.13.4"
Pillow = "^9.1.0"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
